requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Login>=0.6.3,<0.7.0
Flask-SQLAlchemy>=3.1.1,<4.0.0
google-genai>=1.30.0,<2.0.0
